Move error boundry to app level
